+++ src/components/antd-rfh-form-inputs/check-group-input.tsx
@@ -30,7 +30,7 @@ export const CheckGroupInput = <TFieldValues extends FieldValues>({
       name={name}
       control={control}
       render={({ field: { onChange, value } }) => (
-        <Label label={label} tooltip={tooltip}>
+        <Label name={name} label={label} tooltip={tooltip}>
           <Checkbox.Group
             onChange={onChange}
             value={value}