+++ src/components/antd-rfh-form-inputs/text-input.tsx
@@ -40,7 +40,7 @@ export const TextInput = <TFieldValues extends FieldValues>({
       name={name}
       control={control}
       render={({ field: { onChange, value, onBlur } }) => (
-        <Label label={label} tooltip={tooltip}>
+        <Label name={name} label={label} tooltip={tooltip}>
           {(() => {
             switch (type) {
               case 'password':